@dbelcham 2016-11-16
Removed in favour of #4300